# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.FieldRequest:
    properties:
      collection_id:
        type: integer
      data_type_id:
        type: integer
      group_id:
        type: integer
      label:
        type: string
      status:
        type: boolean
    required:
    - data_type_id
    - label
    type: object
  handlers.FormFieldReference:
    properties:
      field_row:
        type: integer
      field_span:
        type: integer
      fields_id:
        type: integer
      validations:
        type: string
    required:
    - fields_id
    type: object
  handlers.FormFieldRequest:
    properties:
      field_id:
        type: integer
      field_row:
        type: integer
      field_span:
        type: integer
      form_group_id:
        type: integer
      form_id:
        type: integer
      validation:
        type: string
    required:
    - field_id
    - form_id
    type: object
...
  handlers.FormRequest:
    properties:
      data_type_id:
        type: integer
      description:
        type: string
      fields:
        items:
          $ref: '#/definitions/handlers.FormFieldReference'
        type: array
      form_name:
        type: string
      service_id:
        type: integer
      status:
        type: boolean
    required:
    - data_type_id
    - form_name
    type: object
  handlers.GroupRequest:
    properties:
      group_name:
        type: string
    required:
    - group_name
    type: object
...
  structs.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      status:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the Kora API server for form management.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kora API
  version: "1.0"
paths:
...
  /field:
    post:
      consumes:
      - application/json
      description: Create a new field
      parameters:
      - description: Field Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.FieldRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Create field
      tags:
      - fields
  /field/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a field by its ID
      parameters:
      - description: Field ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Delete field
      tags:
      - fields
    get:
      consumes:
      - application/json
      description: Retrieve a field by its ID
      parameters:
      - description: Field ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Get field
      tags:
      - fields
    put:
      consumes:
      - application/json
      description: Update an existing field by its ID
      parameters:
      - description: Field ID
        in: path
        name: id
        required: true
        type: integer
      - description: Field Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.FieldRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Update field
      tags:
      - fields
  /form:
    post:
      consumes:
      - application/json
      description: Create a new form with fields
      parameters:
      - description: Form Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.FormRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Create a new form
      tags:
      - form
  /form/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a form by its ID
      parameters:
      - description: Form ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Delete form
      tags:
      - form
    get:
      consumes:
      - application/json
      description: Retrieve a form by its ID
      parameters:
      - description: Form ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Get form by ID
      tags:
      - form
    put:
      consumes:
      - application/json
      description: Update an existing form by its ID
      parameters:
      - description: Form ID
        in: path
        name: id
        required: true
        type: integer
      - description: Form Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.FormRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Update form
      tags:
      - form
  /form_fields:
    post:
      consumes:
      - application/json
      description: Create a form field association
      parameters:
      - description: Form Field Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.FormFieldRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Create form field
      tags:
      - form-fields
  /form_fields/multiple:
    post:
      consumes:
      - application/json
      description: Create multiple form field associations
      parameters:
      - description: Form Field Requests
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/handlers.FormFieldRequest'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Create multiple form fields
      tags:
      - form-fields
...
  /groups:
    get:
      consumes:
      - application/json
      description: Retrieve all groups
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Get all groups
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Create a new group
      parameters:
      - description: Group Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.GroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Create group
      tags:
      - groups
  /groups/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a group by its ID
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Delete group
      tags:
      - groups
    get:
      consumes:
      - application/json
      description: Retrieve a group by its ID
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Get group
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: Update an existing group by its ID
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Group Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.GroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Update group
      tags:
      - groups
...
schemes:
- http
- https
swagger: "2.0"